backend/requirements.txt
asgiref==3.7.2
astroid==3.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
coverage==7.4.1
cycler==0.12.1
dill==0.3.8
Django~=4.2.10
django-cors-headers==4.3.1
django-jenkins==0.110.0
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==7.0.0
fonttools==4.47.2
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
isort==5.13.2
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pep8==1.7.1
pillow==10.2.0
platformdirs==4.2.0
pyarrow==15.0.0
pycodestyle==2.11.1
pyflakes==3.2.0
pylint==3.0.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.3
types-requests==2.31.0.20240125
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
